requirements.txt
# Core Hugging Face packages
transformers>=4.40.0
torch>=2.0.0
tokenizers>=0.15.0
datasets>=2.18.0
accelerate>=0.27.0

# Additional useful packages for LLM work
sentencepiece>=0.1.99
protobuf>=3.20.0
safetensors>=0.4.0

# Optional but recommended packages
gradio>=4.0.0  # For building web interfaces
wandb>=0.16.0  # For experiment tracking
evaluate>=0.4.0  # For model evaluation metrics
peft>=0.8.0  # Parameter Efficient Fine-Tuning

# Development and utility packages
jupyter>=1.0.0
ipywidgets>=8.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# Optional: For working with specific model types
bitsandbytes>=0.42.0  # For quantization
# flash-attn>=2.5.0  # For flash attention (commented out - requires compilation)